+++ radius/src/core/rfc2868.rs
@@ -74,7 +74,7 @@ pub fn delete_tunnel_type(packet: &mut Packet) {
 }
 /// Add `tunnel_type` tagged value-defined integer value to a packet.
 pub fn add_tunnel_type(packet: &mut Packet, tag: Option<&Tag>, value: TunnelType) {
-    packet.add(AVP::from_tagged_u32(TUNNEL_TYPE_TYPE, tag, value as u32));
+    packet.add(AVP::from_tagged_u32(TUNNEL_TYPE_TYPE, tag, value));
 }
 /// Lookup a `tunnel_type` tagged value-defined integer value from a packet.
 ///
@@ -102,11 +102,7 @@ pub fn delete_tunnel_medium_type(packet: &mut Packet) {
 }
 /// Add `tunnel_medium_type` tagged value-defined integer value to a packet.
 pub fn add_tunnel_medium_type(packet: &mut Packet, tag: Option<&Tag>, value: TunnelMediumType) {
-    packet.add(AVP::from_tagged_u32(
-        TUNNEL_MEDIUM_TYPE_TYPE,
-        tag,
-        value as u32,
-    ));
+    packet.add(AVP::from_tagged_u32(TUNNEL_MEDIUM_TYPE_TYPE, tag, value));
 }
 /// Lookup a `tunnel_medium_type` tagged value-defined integer value from a packet.
 ///
